Refactor shared utilities layout
