+++ src/shared/ast/index.js
@@ -1,0 +1,4 @@
+export * from "./comments.js";
+export * from "./location-keys.js";
+export * from "./locations.js";
+export * from "./node-helpers.js";
