+++ src/shared/ast/node-helpers.js
@@ -1,0 +1,345 @@
+import { asArray, isNonEmptyArray } from "../utils/array.js";
+
+// Shared AST helper utilities focused on querying common node shapes.
+// Centralizes frequently repeated guards so printer and transform modules
+// can reuse the same defensive checks without duplicating logic.
+
+/**
+ * Retrieve the sole declarator from a variable declaration node.
+ *
+ * @param {object | null | undefined} node - Potential variable declaration
+ *     node to inspect.
+ * @returns {object | null} The single declarator when present, otherwise
+ *     `null`.
+ */
+function getSingleVariableDeclarator(node) {
+    if (!node || node.type !== "VariableDeclaration") {
+        return null;
+    }
+
+    const { declarations } = node;
+    if (!Array.isArray(declarations) || declarations.length !== 1) {
+        return null;
+    }
+
+    const [declarator] = declarations;
+    if (!declarator || declarator.type !== "VariableDeclarator") {
+        return null;
+    }
+
+    return declarator;
+}
+
+/**
+ * Read and normalize the `kind` field from a variable declaration node.
+ *
+ * @param {object | null | undefined} node - Possible variable declaration
+ *     wrapper exposed by the parser.
+ * @returns {"var" | "global" | "static" | string | null} Lowercase
+ *     declaration keyword when present, or `null` when the field is
+ *     missing/unknown. The return type intentionally remains permissive so the
+ *     printer can surface new keywords added by the parser without needing a
+ *     project-wide update.
+ */
+function getVariableDeclarationKind(node) {
+    if (!node || node.type !== "VariableDeclaration") {
+        return null;
+    }
+
+    const { kind } = node;
+    if (typeof kind !== "string" || kind.length === 0) {
+        return null;
+    }
+
+    return kind.toLowerCase();
+}
+
+/**
+ * Compare a declaration node against a specific keyword.
+ *
+ * @param {object | null | undefined} node - Candidate variable declaration.
+ * @param {string | null | undefined} expectedKind - Keyword to match (e.g.
+ *     `"var"`). The comparison is case-insensitive so callers may pass
+ *     user input without pre-normalizing it.
+ * @returns {boolean} `true` when `node.kind` resolves to the
+ *     provided keyword.
+ */
+function isVariableDeclarationOfKind(node, expectedKind) {
+    if (typeof expectedKind !== "string" || expectedKind.length === 0) {
+        return false;
+    }
+
+    const normalizedKind = getVariableDeclarationKind(node);
+    if (normalizedKind === null) {
+        return false;
+    }
+
+    return normalizedKind === expectedKind.toLowerCase();
+}
+
+function isVarVariableDeclaration(node) {
+    return isVariableDeclarationOfKind(node, "var");
+}
+
+/**
+ * Normalize various identifier-like nodes to a comparable string.
+ *
+ * @param {string | null | undefined | { type?: string, name?: unknown, value?: unknown, object?: unknown, property?: unknown }} node
+ *     Any AST fragment that may carry a name. String values are returned as-is.
+ * @returns {string | null} Canonical identifier text, using underscores to
+ *     flatten member access (e.g. `foo.bar` -> `"foo_bar"`) or
+ *     `null` when the node does not resolve to a string name. The helper
+ *     treats unexpected node shapes defensively, which allows callers inside
+ *     hot printer paths to skip type checks without risking runtime failures.
+ */
+function getIdentifierText(node) {
+    if (node == undefined) {
+        return null;
+    }
+
+    if (typeof node === "string") {
+        return node;
+    }
+
+    // Hoist the common type lookup so the switch below can reuse it without
+    // repeatedly touching the same field during hot traversal paths.
+    const { type } = node;
+
+    switch (type) {
+        case "Identifier": {
+            const { name } = node;
+            return typeof name === "string" ? name : null;
+        }
+        case "Literal": {
+            const { value } = node;
+            return typeof value === "string" ? value : null;
+        }
+        case "MemberDotExpression": {
+            const { object, property } = node;
+
+            if (
+                !object ||
+                object.type !== "Identifier" ||
+                !property ||
+                property.type !== "Identifier"
+            ) {
+                return null;
+            }
+
+            // String concatenation avoids the template literal machinery in this
+            // hot branch, shaving work inside tight printer loops.
+            return object.name + "_" + property.name;
+        }
+        case "MemberIndexExpression": {
+            const object = node.object;
+
+            if (!object || object.type !== "Identifier") {
+                return null;
+            }
+
+            const property = node.property;
+            if (!Array.isArray(property) || property.length !== 1) {
+                return null;
+            }
+
+            const indexText = getMemberIndexText(property[0]);
+            return indexText == undefined
+                ? null
+                : object.name + "_" + indexText;
+        }
+        default: {
+            const { name } = node;
+            return typeof name === "string" ? name : null;
+        }
+    }
+}
+
+/**
+ * Extract the printable index portion of a {@link MemberIndexExpression}.
+ *
+ * @param {string | null | undefined | object} indexNode Possible node nested
+ *     within `MemberIndexExpression.property`. Arrays are handled by the
+ *     caller; this helper focuses on the single item case enforced by the
+ *     parser.
+ * @returns {string | null} Resolved index name or `null` when the parser
+ *     emitted a non-string structure (for example, computed expressions). The
+ *     defensive guards let callers gracefully skip edge cases without
+ *     introducing conditional branches at the call site.
+ */
+function getMemberIndexText(indexNode) {
+    if (typeof indexNode === "string") {
+        return indexNode;
+    }
+
+    if (indexNode == undefined) {
+        return null;
+    }
+
+    const indexName = indexNode.name;
+    if (typeof indexName === "string") {
+        return indexName;
+    }
+
+    const indexType = indexNode.type;
+    if (indexType === "Literal") {
+        const value = indexNode.value;
+        return typeof value === "string" ? value : null;
+    }
+
+    return getIdentifierText(indexNode);
+}
+
+/**
+ * Return the sole property entry from a {@link MemberIndexExpression} when the
+ * parser emitted exactly one index element. Several transforms guard against
+ * unexpected array shapes before inspecting the property, so this helper
+ * centralizes the defensive checks and keeps those call sites in sync.
+ *
+ * @param {unknown} node Candidate member index expression.
+ * @returns {unknown | null} The single property entry or `null` when missing.
+ */
+function getSingleMemberIndexPropertyEntry(node) {
+    if (!isNode(node) || node.type !== "MemberIndexExpression") {
+        return null;
+    }
+
+    const { property } = node;
+    if (!Array.isArray(property) || property.length !== 1) {
+        return null;
+    }
+
+    const [propertyEntry] = property;
+    return propertyEntry ?? null;
+}
+
+/**
+ * Safely read the argument array from a call-like AST node.
+ *
+ * @param {object | null | undefined} callExpression Potential call expression
+ *     node that may expose an `arguments` array.
+ * @returns {Array<unknown>} Normalized argument collection. Returns a shared
+ *     empty array when no arguments exist so callers can iterate without
+ *     additional null checks.
+ */
+// Delegate to the shared array normaliser so call-expression traversals always
+// reuse the same frozen empty array rather than recreating bespoke helpers.
+function getCallExpressionArguments(callExpression) {
+    if (!isNode(callExpression)) {
+        return asArray();
+    }
+
+    return asArray(callExpression.arguments);
+}
+
+function getArrayProperty(node, propertyName) {
+    if (!isNode(node)) {
+        return [];
+    }
+
+    if (typeof propertyName !== "string" || propertyName.length === 0) {
+        return [];
+    }
+
+    return asArray(node[propertyName]);
+}
+
+function hasArrayPropertyEntries(node, propertyName) {
+    if (!isNode(node)) {
+        return false;
+    }
+
+    if (typeof propertyName !== "string" || propertyName.length === 0) {
+        return false;
+    }
+
+    return isNonEmptyArray(node[propertyName]);
+}
+
+function getBodyStatements(node) {
+    return getArrayProperty(node, "body");
+}
+
+function hasBodyStatements(node) {
+    return hasArrayPropertyEntries(node, "body");
+}
+
+function isProgramOrBlockStatement(node) {
+    if (!isNode(node)) {
+        return false;
+    }
+
+    const { type } = node;
+    if (typeof type !== "string") {
+        return false;
+    }
+
+    return type === "Program" || type === "BlockStatement";
+}
+
+function getLiteralStringValue(node) {
+    if (!isNode(node) || node.type !== "Literal") {
+        return null;
+    }
+
+    const { value } = node;
+    if (typeof value !== "string") {
+        return null;
+    }
+
+    return value.toLowerCase();
+}
+
+function getBooleanLiteralValue(node, options = {}) {
+    if (!isNode(node) || node.type !== "Literal") {
+        return null;
+    }
+
+    const acceptBooleanPrimitives =
+        typeof options === "boolean"
+            ? options
+            : !!options?.acceptBooleanPrimitives;
+
+    const { value } = node;
+
+    if (value === true || value === false) {
+        if (!acceptBooleanPrimitives) {
+            return null;
+        }
+
+        return value ? "true" : "false";
+    }
+
+    const normalized = getLiteralStringValue(node);
+    return normalized === "true" || normalized === "false" ? normalized : null;
+}
+
+function isBooleanLiteral(node, options) {
+    return getBooleanLiteralValue(node, options) !== null;
+}
+
+function isUndefinedLiteral(node) {
+    return getLiteralStringValue(node) === "undefined";
+}
+
+function isNode(value) {
+    return value != undefined && typeof value === "object";
+}
+
+export {
+    getSingleVariableDeclarator,
+    getVariableDeclarationKind,
+    getIdentifierText,
+    getCallExpressionArguments,
+    getArrayProperty,
+    hasArrayPropertyEntries,
+    getBodyStatements,
+    hasBodyStatements,
+    isProgramOrBlockStatement,
+    getSingleMemberIndexPropertyEntry,
+    getBooleanLiteralValue,
+    isBooleanLiteral,
+    isUndefinedLiteral,
+    isNode,
+    isVariableDeclarationOfKind,
+    isVarVariableDeclaration
+};
